% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenbgs.R
\name{eigenbgs}
\alias{eigenbgs}
\title{Eigen background subtraction algorithm}
\usage{
eigenbgs(b, nf, thresh)
}
\arguments{
\item{b}{3-D matrix containing grayscaled video frames.}

\item{nf}{Number of previous frames to be considered. Its value has to be greater than 1.}

\item{thresh}{threshold required to obtain foreground images.value of thresh can be set to 10 - 15 approximately.}
}
\value{
A 3-D matrix of frames containing foreground obtained after background subtraction
 is applied(binary images).
}
\description{
This function uses previous frames to construct a background model in eigen space. This eigen background
model is then subtracted from current frame to obtain the foreground images.
}
