% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{guassianbgs}
\alias{guassianbgs}
\title{Gaussian background subtraction algorithm}
\usage{
guassianbgs(b, thresh)
}
\arguments{
\item{b}{3-D matrix containing grayscaled video frames.}

\item{thresh}{threshold required to obtain foreground images. Its value can be set around 3.5 - 4.5
depending upon the accuracy.}
}
\value{
A 3-D matrix of frames containing foreground obtained after background subtraction
 is applied(binary images).
}
\description{
This functions creates a gaussian background model using the previous grayscale frames.
gaussian parameters (i.e mean and variance) are updated after every new frame is encountered.
}
