% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbgs_dependency.R
\name{inst_xug}
\alias{inst_xug}
\title{This function helps to add the Xuggle-5.4 to classpath.}
\usage{
inst_xug()
}
\description{
This function asks the user whether to download and add xuggle-5.4 to classpath. Based upon the
input provided by the user, specific action is taken
}
