% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoreading.R
\name{readvideo}
\alias{readvideo}
\title{Read a Video}
\usage{
readvideo(videoURL)
}
\arguments{
\item{videoURL}{Path to the input video file}
}
\value{
A matrix of the grayscale frames.
}
\description{
This function takes as input URL of the video. It can return maximum hundred frames of the video
in which each frame is resized to 100 X 100 pixel image.
If the video contains less than hundred frames. then it will return all the frames otherwise it
will automatically return first hundred frames. Frames returned are grayscale frames.
}
\examples{
##Save the URL of the video file into R session and then load videoframes
videoURL <- system.file("extdata","daria_skip.avi",package = "Rbgs")
frames <- readvideo(videoURL)
}
