% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framevideoread.R
\name{readvideoframe}
\alias{readvideoframe}
\title{Read the specified frames of a video.}
\usage{
readvideoframe(videoURL, start, end)
}
\arguments{
\item{videoURL}{Path to the input video file}

\item{start}{It indicates the first frame you wish to read.IT should be in range zero to frame count
and smaller than end parameter.}

\item{end}{It indicates the last frame you wish to read. It should be greater than start and should
be in range 0 to frame count.}
}
\value{
A matrix of the grayscale frames.
}
\description{
This function takes as input URL of the video. It reads only those frames of the video that
have been specified by the user. It requires two more parameters 'start' and 'end', that defines the
range of the videoframes to be retrieved.
}
\examples{
##Save the URL of the video file into R session and then load the required videoframes
videoURL <- system.file("extdata","jog.mp4",package = "Rbgs")
frames <- readvideoframe(videoURL,90,110)
}
