% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framecount.R
\name{totalframe}
\alias{totalframe}
\title{Total number of frames in a video.}
\usage{
totalframe(videoURL)
}
\arguments{
\item{videoURL}{Path to the input video file}
}
\value{
count total no of frames in a video.
}
\description{
This function takes as input URL of the video. This function returns total number of frames contained in
the video file.
}
