% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varbgs.R
\name{varbgs}
\alias{varbgs}
\title{Variance based background subtraction algorithm}
\usage{
varbgs(b, nf, thresh)
}
\arguments{
\item{b}{3-D matrix containing grayscaled video frames.}

\item{nf}{number of frames to be considered to construct background model. Its value has to be greater
greater than 1. Smaller value gives better results.}

\item{thresh}{threshold required to obtain foreground images.Its value can be around 220.}
}
\value{
A 3-D matrix of frames containing foreground obtained after background subtraction
 is applied(binary images).
}
\description{
This function calculates the variance of previous nf number of frames and obtains the foreground by
subtracting the variance obtained from the current frame.
}
