% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtmean.R
\name{wtmean}
\alias{wtmean}
\title{Weighted Mean background subtraction algorithm}
\usage{
wtmean(b, nf, thresh)
}
\arguments{
\item{b}{3-D matrix containing grayscaled video frames.}

\item{nf}{It is size of buffer that contains previous frames and its value can be set to three only.}

\item{thresh}{threshold required to obtain foreground images.Its value can be around 10 - 30.}
}
\value{
A 3-D matrix of frames containing foreground obtained after background subtraction
 is applied(binary images).
}
\description{
This function calculates the weighted mean of previous three frames. The buffer maintained
here can contain only three frames.
}
