% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtvariance.R
\name{wtvar}
\alias{wtvar}
\title{Weighted variance background subtraction algorithm}
\usage{
wtvar(b, nf, thresh)
}
\arguments{
\item{b}{3-D matrix containing grayscaled video frames.}

\item{nf}{previous number of frames to be considered to build background.}

\item{thresh}{threshold required to obtain foreground images.Its value can be set to 180-230 or more approximately.}
}
\value{
A 3-D matrix of frames containing foreground obtained after background subtraction
 is applied(binary images).
}
\description{
This function calculates the weighted variance of previous three frames. The buffer maintained
here can contain only three frames.
}
